requirements.txt
autopep8==2.0.1
bert-extractive-summarizer==0.10.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
cymem==2.0.7
Cython==0.29.32
DAWG-Python==0.7.2
docopt==0.6.2
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
huggingface-hub==0.11.1
idna==3.4
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.1
miniful==0.0.6
murmurhash==1.0.9
networkx==2.8.8
nlp-rake==0.0.2
nltk==3.8.1
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pathy==0.10.1
preshed==3.0.8
pycodestyle==2.10.0
pydantic==1.10.4
pyFUME==0.2.25
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymystem3==0.2.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
rake-nltk==1.0.6
regex==2022.10.31
requests==2.28.1
scapy==2.5.0
scikit-learn==1.2.0
scipy==1.10.0
segtok==1.5.11
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
tabulate==0.9.0
thinc==8.1.6
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
Werkzeug==2.2.2
yake==0.4.8
